[shop][fix] 파일 삭제
