+++ market-shop/src/test/java/com/pmc/market/controller/ShopControllerTest.java
@@ -1,17 +1,72 @@
 package com.pmc.market.controller;
 
+import com.pmc.market.ShopApplication;
+import com.pmc.market.entity.Shop;
+import com.pmc.market.service.ShopService;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.http.MediaType;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 
-import static org.junit.jupiter.api.Assertions.*;
+import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@AutoConfigureWebTestClient
-class ShopControllerTest {
+import static org.hamcrest.collection.IsCollectionWithSize.hasSize;
+import static org.mockito.Mockito.when;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+
+@ExtendWith(SpringExtension.class)
+@SpringBootTest(classes = {ShopApplication.class}, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@AutoConfigureMockMvc
+public class ShopControllerTest {
+
+    @Autowired
+    MockMvc mockMvc;
+    @MockBean
+    private ShopService shopService;
 
     @Test
-    void Shop_목록_조회(){
+    void 쇼핑몰_목록을_가져온다() throws Exception {
+        List<Shop> shops = new ArrayList<>();
+        shops.add(Shop.builder()
+                .id(1L)
+                .name("쇼핑몰1")
+                .telephone("010-0000-0000")
+                .businessName("쇼핑몰1")
+                .fullDescription("쇼핑몰 설명")
+                .owner("주인")
+                .shortDescription("악세사리 쇼핑몰")
+                .regDate(LocalDateTime.now())
+                .period(LocalDateTime.now().plusYears(1))
+                .businessNumber("00-000-000")
+                .build());
+        shops.add(Shop.builder()
+                .id(2L)
+                .name("쇼핑몰2")
+                .telephone("010-0000-0000")
+                .businessName("쇼핑몰2")
+                .fullDescription("쇼핑몰 설명2")
+                .owner("주인2")
+                .shortDescription("핸드메이드 쇼핑몰")
+                .regDate(LocalDateTime.now())
+                .period(LocalDateTime.now().plusYears(1))
+                .businessNumber("00-000-000")
+                .build());
+        when(shopService.findAll()).thenReturn(shops);
 
+        mockMvc.perform(MockMvcRequestBuilders.get("/shops")
+                .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.data", hasSize(2)))
+                .andDo(print());
     }
 }
